requirements.txt
streamlit
openai
moviepy
srt
ffmpeg-python
python-dotenv
faster-whisper
pandas
